+++ android/src/main/java/com/plugin/flutter/zsdk/PrinterUtils.java
@@ -15,22 +15,13 @@ class PrinterUtils
     * Reboots the printer.
     * Returns true if successfully rebooted, false otherwise
     */
-   public static boolean reboot(Connection connection) {
-      try {
-         if(!connection.isConnected()) connection.open();
-         // Another way would be using the following ZPL command:
-         // connection.write("^XA^JUS^XZ".getBytes());
-         ZebraPrinter printer = ZebraPrinterFactory.getInstance(connection);
-         printer.reset();
-         connection.close();
-         return true;
-      } catch (ConnectionException e) {
-         e.printStackTrace();
-      } catch (ZebraPrinterLanguageUnknownException e1) {
-         e1.printStackTrace();
-      } catch(Exception e) {
-         e.printStackTrace();
-      }
-      return false;
+   public static boolean reboot(Connection connection) throws Exception {
+      if(!connection.isConnected()) connection.open();
+      // Another way would be using the following ZPL command:
+      // connection.write("^XA^JUS^XZ".getBytes());
+      ZebraPrinter printer = ZebraPrinterFactory.getInstance(connection);
+      printer.reset();
+      connection.close();
+      return true;
    }
 }