+++ android/src/main/java/com/plugin/flutter/zsdk/VirtualDeviceUtils.java
@@ -17,7 +17,7 @@ class VirtualDeviceUtils {
     * This changes the printer Virtual Device only if the current Virtual Device on the printer is different that the one to be changed to and reboots the printer for the changes to take effect.
     * Returns true if the change was necessary and applied, false otherwise.
     */
-   public static boolean changeVirtualDevice(Connection connection, String virtualDevice) throws ConnectionException {
+   public static boolean changeVirtualDevice(Connection connection, String virtualDevice) throws Exception {
       if(TextUtils.isEmpty(virtualDevice)) return false;
       if(connection == null) return false;
       if(!connection.isConnected()) connection.open();